split component and extract types
