+++ frontend/app/src/entities/nodes/convert/types.ts
@@ -1,5 +1,8 @@
 import type { EmptyFieldValue, FormFieldValue } from "@/shared/components/form/type";
 
+import type { NodeCore, NodeObject } from "@/entities/nodes/types";
+import type { AttributeSchema, ModelSchema } from "@/entities/schema/types";
+
 export interface ConvertFieldMapping {
   is_mandatory: boolean;
   source_field_name: string | null;
@@ -17,3 +20,42 @@ export type ConvertFormFieldValue =
       value: FormFieldValue["value"];
     }
   | EmptyFieldValue;
+
+export interface ConvertSourceInputProps {
+  sourceObject: NodeObject;
+  sourceSchema: ModelSchema;
+  mapping?: ConvertFieldMapping;
+  className?: string;
+  value: ConvertFormFieldValue;
+  onChange: (value: ConvertFormFieldValue) => void;
+}
+
+export interface ConvertSourceOption {
+  source: {
+    type: "source";
+    name: string;
+    label: string | null | undefined;
+  };
+  isDefaultMatch: boolean;
+}
+
+export interface AttributeSourceOption extends ConvertSourceOption {
+  value: string | string[] | number | boolean | null;
+  label: React.ReactNode;
+}
+
+export interface RelationshipOneSourceOption extends ConvertSourceOption {
+  value: NodeCore | null;
+}
+
+export interface RelationshipManySourceOption extends ConvertSourceOption {
+  value: Array<NodeCore> | null;
+}
+
+export interface ConvertSourceAttributeComboboxProps extends ConvertSourceInputProps {
+  attribute: AttributeSchema;
+}
+
+export interface ConvertSourceRelationshipInputProps extends ConvertSourceInputProps {
+  peer: string;
+}
